fix remarks after 1 attempt
