+++ ErshovVladislav/Task1/README.md
@@ -1,16 +1,34 @@
 ## Как использовать
 
-- Запустить из папки `src/main/java/QuickSortUsingSerialSetOfSamples`
+#### Установка MPJ Express
+
+- Скачать MPJ Express https://sourceforge.net/projects/mpjexpress/files/releases/mpj-v0_44.zip/download
+- Распаковать его
+- Добавить корень распакованного архива в `MPJ_HOME`
+- Добавить папку `bin` из корня распакованного архива в `Path`
+
+#### Использование
+
+- Запустить из папки `src`
 ```
-javac -cp .;%MPJ_HOME%/lib/mpj.jar QuickSortUsingSerialSetOfSamples/Main.java
+javac -cp .;%MPJ_HOME%/lib/mpj.jar main/java/quicksort/Main.java
 ```
 где `MPJ_HOME` это путь до MPJ Express
 
 - Затем запустить
 ```
-mpjrun.bat -np n QuickSortUsingSerialSetOfSamples.Main
+mpjrun.bat -np n main.java.quicksort.Main path/to/input/file path/to/output/file
 ```
-где `n` это количество вычислительных узлов
+где:
+
+- `n` это количество вычислительных узлов
+
+- `path/to/input/file` это путь от `src` до файла с изначальным массивом
+- `path/to/output/file` это путь от `src` до файла, куда запишется отсортированный массив
+
+
+
+В корне проекта в папке `examples` есть примеры начальных массивов
 
 ## Использование с maven
 
@@ -23,14 +41,14 @@ mvn -v
 - Если нет, то исправить это:
 - Скачать архив https://dlcdn.apache.org/maven/maven-3/3.9.0/binaries/apache-maven-3.9.0-bin.zip
 - Распаковать его
-- Добавить папку bin из корня распакованного архива в Path
+- Добавить папку `bin` из корня распакованного архива в `Path`
 
 #### Необходимые зависимости из pom.xml
 
-- java version 17
+- `java` version 17
 
-- mpj version 0.44
-- junit4 version 4.13.2
+- `mpj` version 0.44
+- `junit4` version 4.13.2
 
 #### Использование
