+++ ErshovVladislav/Task1/examples/initialArraySmall.txt
@@ -1,0 +1,1 @@
+54 21 -3 1 8 3
