+++ ErshovVladislav/Task1/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.springframework</groupId>
-    <artifactId>gs-maven</artifactId>
+    <artifactId>quicksort-maven</artifactId>
     <packaging>jar</packaging>
     <version>0.1.0</version>
 
@@ -27,6 +27,12 @@
             <version>4.13.2</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+            <version>4.2.0</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
@@ -45,7 +51,7 @@
                             <transformers>
                                 <transformer
                                     implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>QuickSortUsingSerialSetOfSamples.Main</mainClass>
+                                    <mainClass>main.java.quicksort.Main</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
