--- ErshovVladislav/Task1/resources/initialArraySmall.txt
@@ -1,1 +1,0 @@
-54 21 -3 1 8 3
